Moved MutationValidator to configurations package
